run all regions, and do not stop on first failure

+++ bin/garmin-all.sh
@@ -48,6 +48,7 @@ download_region ()
   mv -f $tmp $sub_region.osm.pbf
 }
 
+exit_status=0
 for region in $garmin_regions
 do
   $debug && echo "region=$region format=$garmin_formats"
@@ -57,11 +58,19 @@ do
   (
     mkdir -p $continent
     cd $continent
-    download_region $region $sub_region
-    env osm2xxx_max_jobs="8" OSM_CHECKSUM=false pbf2osm_max_cpu_time=72000 max_file_size_garmin=59950000 \
-      BBBIKE_TMPFS=/tmp \
-        nice -n $nice_level $time $HOME/projects/bbbike/world/bin/pbf2osm --garmin-${garmin_formats} $sub_region.osm.pbf $region
-    rm -f $sub_region.osm.pbf
+    if download_region $region $sub_region; then
+      env osm2xxx_max_jobs="8" OSM_CHECKSUM=false pbf2osm_max_cpu_time=72000 max_file_size_garmin=59950000 \
+        BBBIKE_TMPFS=/tmp \
+          nice -n $nice_level $time $HOME/projects/bbbike/world/bin/pbf2osm --garmin-${garmin_formats} $sub_region.osm.pbf $region || exit_status=1
+      rm -f $sub_region.osm.pbf
+    else
+      echo "could not download $url - skip"
+      exit_status=2
+    fi
   )
 done
+
+exit $exit_status
+
+#EOF
